Add more changes

+++ UI/js/allgroups.js
@@ -1,10 +1,5 @@
-// token = localStorage.removeItem('token');
-//     if (token == null) {
-//         alert('please login');
-//         window.location.replace('UI/index.html');
-//     };
 function myGroups() {
-    let recieveUrl = 'http://127.0.0.1:5000/api/v2/groups';
+    let recieveUrl = 'https://epiks.herokuapp.com/api/v2/groups';
     token = localStorage.getItem('token');
     if (token === null) {
         alert('please login');
@@ -25,25 +20,22 @@ function myGroups() {
                 var d = response.data;
                 console.log(d)
                 let group = `<table>
-                    <tr>
-                        
-                        <th>Name</th> 
-                        
-                        <th>Role</th>
-                        <th>DELETE</th>
-                        <th>Add users to group</th>
-                    </tr>
+
                         `;
                 d.forEach((grp) => {
 
                     group +=
                         `  
                         <tr>
+                        <td><input type = "radio" value = "${grp.id}" name = "identity"onchange = "myFunctions(${grp.id})"></td>
+
                         <td>${grp.name}</td>
                         
                         <td>${grp.role}</td>
                         <td onclick='deletes(${grp.id})'><img src="../images/delete.png" alt="delete" height="22" width="22"></td>
-                        <td><a href = add_members_to_group.html(${grp.id})>addmembers</a></td>
+                        <td><a href = "../html/add_members_to_group.html?idi=${grp.id}" id = "detail">addmembers</a></td>
+                        <td><a href = "../html/view_groupmembers.html?idi=${grp.id}" id = "detail">view group members</a></td>
+
                         <td onclick='add()'><button>Add</button></td></tr>
                         `
                 });
@@ -53,61 +45,47 @@ function myGroups() {
         })
 }
 
+function myFunctions(id) {
+    document.getElementById("ikons").style.display = "block";
+    console.log(id)
+    localStorage.setItem('identity', id)
+    document.getElementById("ikons").style.display = "block";
+    return id;
+    // deleteds(id)
+}
+
 function add() {
     let email = document.getElementById('newmember').value;
     let data = {
         "email": email
     }
-    
-        let readmsgUrl = `http://127.0.0.1:5000/api/v2/groups/41/users`;
-        token = localStorage.getItem('token');
-        fetch(readmsgUrl, {
-                method: 'POST',
-                // mode: "no-cors",
-   
-                headers: {
-                    'Authorization': `Bearer ${token}`
-                }
-            })
-            .then(res => res.json())
-            .then(response => {
-                console.log(response)
-                // if (response.error === "you have no recieved messages") {
-                //     document.getElementById('no_mail').innerHTML = response.error;
-                // } else if (response.status == 200) {
-                //     var d = response.data;
-                //     console.log(d)
-                //     window.location.replace('./user_dash.html')
-                // }
-            })
-    }
 
+    let readmsgUrl = `https://epiks.herokuapp.com/api/v2/groups/41/users`;
+    token = localStorage.getItem('token');
+    fetch(readmsgUrl, {
+            method: 'POST',
+            mode: "no-cors",
 
-   
-    //                 if (data.message === 'thanks for registering with Epic mail') {
-    //                     token = data.data[0].token;
-    //                     localStorage.setItem('token', token)
+            headers: {
+                'Authorization': `Bearer ${token}`
+            }
+        })
+        .then(res => res.json())
+        .then(response => {
+            console.log(response)
 
-    //                     window.location.replace('./user_dash.html')
+        })
+}
 
-    //                 } else {
-
-    //                     document.getElementById("negativeresponse").innerHTML = `<h1 style="color:red; background-color:pink; margin: 2%; font-size: 100%;" >${data.error}</h1>`
-    //                     // document.getElementById("negativeresponse").style.display = block
-    //                 }
-    //             })
-    //     }
-    // }
-
-
-
-
-
-    function deletes(id) {
-        let readmsgUrl = `http://127.0.0.1:5000/api/v2/groupss/${id}`;
+function deletes() {
+    id = localStorage.getItem('identity');
+    console.log(id)
+    if (confirm("Are you sure you want to delete this group!")) {
+        let readmsgUrl = `https://epiks.herokuapp.com/api/v2/groupss/${id}`;
         token = localStorage.getItem('token');
         fetch(readmsgUrl, {
                 method: 'DELETE',
+                mode: 'cors',
 
                 headers: {
                     'Authorization': `Bearer ${token}`
@@ -126,28 +104,4 @@ function add() {
                 }
             })
     }
-    // function deletes(id) {
-    //     if (confirm("Are you sure you want to delete this group!")){
-    //     let readmsgUrl = `https://epiks.herokuapp.com/api/v2/groupss/${id}`;
-    //     token = localStorage.getItem('token');
-    //     fetch(readmsgUrl, {
-    //             method: 'DELETE',
-
-    //             headers: {
-    //                 'Authorization': `Bearer ${token}`
-    //             }
-    //         })
-    //         .then(res => res.json())
-    //         .then(response => {
-    //             console.log(response)
-    //             if (response.error === "you have no groups") {
-    //                 document.getElementById('no_mail').innerHTML = response.error;
-    //             } else if (response.status == 200) {
-    //                 var d = response.data;
-    //                 console.log(d)
-    //                 document.getElementById('no_mail').innerHTML = d;
-    //                 window.location.replace('./view_groups.html')
-    //             }
-    //         })
-    // }
-    // }
+}
